Update Codecademy_Hurricane_Analysis.py
Fixed it! Convert damages wasn't returning anything from the function call, which was breaking the dict constructor, which in turn was breaking  categorize_by_damage. Everything is working now!

This will probably end this project. Maybe in a few weeks I'll revisit and redo this code as it's super messy now, but for now it accomplishes what the requirements are!